Update actions to use context without client functions
